requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
openai>=1.30.0
Pillow>=10.3.0
python-dotenv>=1.0.0
